+++ options.js
@@ -63,6 +63,8 @@ function createInput(option) { // gestion des différents types d'input
   let inputType = 'input';
   if (['html', 'radio'].includes(option.type)) {
     inputType = 'div';
+  } else if (option.type === 'json') {
+    inputType = 'textarea'; // Utiliser un textarea pour les options de type json
   }
   const input = document.createElement(inputType);
   input.id = option.name;
@@ -77,6 +79,12 @@ function createInput(option) { // gestion des différents types d'input
       case 'text':
         input.type = 'text';
         input.value = optionValue;
+        break;
+      case 'json':
+        input.classList.add('json-input');
+        input.value = displayCategories(optionValue);
+        input.style.height = '300px';
+        input.style.width = '100%';
         break;
       case 'smalltext':
         input.type = 'text';
@@ -123,6 +131,42 @@ async function getOptionValue(option) {
       resolve(valueToReturn);
     });
   });
+}
+
+// Afficher le json sous une forme plus lisible, avec un retour à la ligne après chaque [
+function displayCategories(jsonStr) {
+  let display = '';
+  try {
+    const categories = JSON.parse(jsonStr);
+    categories.forEach(category => {
+      const [name, keywords] = category;
+      display += `${name} : ${keywords.join(' , ')}\n`;
+    });
+  } catch (error) {
+    console.error('Erreur lors de l\'analyse du JSON:', error);
+    // Mettre une erreur pour l'utilisateur
+    alert('Les paramètres pour la gestion des catégories ne sont pas valides, merci de les corriger');
+    display = jsonStr;
+  }
+  console.log(display);
+  return display;
+}
+
+// Récupérer les données affichées et les convertir en JSON
+function getCategoriesFromJsonInput(input) {
+  const categories = [];
+  const lines = input.value.split('\n');
+  lines.forEach(line => {
+    if (line.trim()) { // Vérifier que la ligne n'est pas vide
+      const [name, keywords] = line.split(':');
+      if (name.trim()) { // Vérifier que le nom de la catégorie n'est pas vide
+        const category = [name.trim(), keywords ? keywords.split(',').map(keyword => keyword.trim()) : []];
+        categories.push(category);
+      }
+    }
+  });
+  console.log(JSON.stringify(categories));
+  return categories;
 }
 
 
@@ -245,7 +289,7 @@ chrome.storage.local.get('defaultShortcuts', function (result) {
 
     buttonEvent.target.innerHTML = 'Appuyez sur une touche de fonction ou une combinaison de touches';
     buttonEvent.target.classList.add('modifying');
-    
+
     hotkeys('*', function (event, handler) { // On écoute toutes les pressions de touche
       function saveShortcut(keys) {
         var shortcut = "";
@@ -300,6 +344,8 @@ chrome.storage.local.get(['defaultSettings', 'defaultShortcuts'], function (resu
       let element = document.getElementById(option);
       if (element.classList.contains('radio-group')) {
         valuesToSave[option] = getSelectedRadioValue(option);
+      } else if (element.classList.contains('json-input')) {
+        valuesToSave[option] = JSON.stringify(getCategoriesFromJsonInput(element));
       } else if (element) { // Vérifiez si l'élément existe
         var value = element.type === 'checkbox' ? element.checked : element.value;
         valuesToSave[option] = value;